+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+### 16.0.0
+
+- Switch from axios to fetch
+
+### 15.14.3
+
+- Add `priority` field to promotion response
+
+### 15.14.2
+
+- Export Promotion Condition types
+
 ### 15.14.1
 
 - Export types